--- admin-sidebar/src/App.css
@@ -1,91 +1,0 @@
-/* App.css */
-.App {
-  display: flex;
-  height: 100vh;
-  margin: 0;
-  padding: 0;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-body {
-  margin: 0;
-  padding: 0;
-}
-
-.Sidebar {
-  height: 100%;
-  width: 300px;
-  background-color: rgb(26, 19, 99);
-}
-
-.SidebarList {
-  height: auto;
-  padding: 0;
-  width: 100%;
-}
-
-.SidebarList .row {
-  width: 100%;
-  height: 60px;
-  
-  list-style-type: none;
-  margin: 0;
-  display: flex;
-  flex-direction: row;
-  color: white;
-  justify-content: center;
-  align-items: center;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-.SidebarList .row:hover {
-  cursor: pointer;
-  background-color: rgb(61, 45, 226);
-}
-
-.SidebarList #active{
-  background-color: rgb(61, 45, 226);
-}
-
-.row #icon {
-  flex: 30%;
-  display: grid;
-  place-items: center;
-}
-
-.row #title {
-  flex: 70%;
-}
-
-.header {
-  background-color: #161616;
-  color: white;
-  padding: 10px 20px;
-  box-sizing: border-box;
-  width: calc(100% - 300px);
-  height: 80px;
-  display: flex;
-  align-items: center;
-  position: fixed;
-  top: 0;
-  left: 300px;
-  font-size: 40px; /* Adjust font size */
-  font-weight: bold; /* Make text bold */
-  font-style: sans-serif;
-}
-
-.main-content {
-  margin-left: 0px;
-  padding-top: 80px;
-  width: calc(100% - 300px);
-  height: calc(100vh - 60px);
-  
-}
-
-.content {
-  flex: 1;
-  display: flex;
-  justify-content: center; /* Center the table horizontally */
-  align-items: center; /* Center the table vertically */
-  padding: 20px;
-}
