--- admin-sidebar/src/Components/Dashboard.js
@@ -1,13 +1,0 @@
-// Components/Dashboard.js
-import React from 'react';
-
-const Dashboard = () => {
-  return (
-    <div>
-      <h1>Dashboard</h1>
-      {/* Add dashboard content here */}
-    </div>
-  );
-}
-
-export default Dashboard;
