--- admin-sidebar/src/Components/Sidebar.js
@@ -1,31 +1,0 @@
-//Sidebar.js
-import React from "react";
-import "../App.css";
-import { SidebarData } from "./SidebarData";
-
-function Sidebar() {
-  return (
-    <div className="Sidebar">
-      <ul className="SidebarList">
-        {SidebarData.map((val, key) => {
-          return (
-            <li
-              key={key}
-              className="row"
-              // eslint-disable-next-line
-              id={window.location.pathname == val.link ? "active" : ""}
-              onClick={() => {
-                window.location.pathname = val.link;
-              }}
-            >
-              <div id="icon">{val.icon}</div>
-              <div id="title">{val.title}</div>
-            </li>
-          );
-        })}
-      </ul>
-    </div>
-  );
-}
-
-export default Sidebar;
