+++ gym-management-systems/home-page/src/components/Equipments.css
