+++ gym-management-systems/home-page/src/components/Sales.css
@@ -1,0 +1,191 @@
+.sales{
+    width: 100%;
+    padding: 20px;
+    box-sizing: border-box;
+    font-family: Arial, Helvetica, sans-serif;
+  }
+  
+  .sales table {
+    width: 100%;
+    border: 1px solid black;
+    border-spacing: 0;
+  }
+  
+  .sales th,
+  .sales td {
+    border: 1px solid #ddd;
+    text-align: center;
+    padding: 10px;
+  }
+  
+  .sales  th {
+    background-color: #f2f2f2;
+    font-size: 18px;
+  }
+  
+  .sales button {
+    margin: 0 5px;
+    padding: 8px;
+    border: none;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+  }
+  
+  .sales button:hover {
+    opacity: 0.8;
+  }
+  
+  .sales .edit-button {
+    background-color: #ff9800;
+    border-radius: 50%;
+  }
+  
+  .sales .delete-button {
+    background-color: #f44336;
+    border-radius: 50%;
+  }
+  
+  .sales .icon {
+    color: white;
+    font-size: 20px;
+  }
+  
+  .table-header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    font-size: 24px;
+    padding: 10px;
+    background-color: #f2f2f2;
+    margin-bottom: 10px;
+    border: 1px solid black;
+    height: 35px;
+  }
+  
+  .add-new-button {
+    display: flex;
+    align-items: center;
+    justify-content: center; 
+    cursor: pointer;
+    color: #fff;
+    background-color: #0d47a1; 
+    padding: 8px 20px; 
+    border-radius: 20px; 
+    font-size: 14px; 
+    font-weight: normal; 
+  }
+  
+  .add-new-button:hover {
+    background-color: #1565c0; 
+  }
+  
+
+  
+  .search-container {
+    display: flex;
+    align-items: center;
+    background-color: #fff;
+    border: 1px solid #ccc;
+    border-radius: 20px;
+    padding: 5px 10px;
+    width: 350px;
+  }
+  
+  .search-icon {
+    color: #757575;
+    margin-right: 5px;
+  }
+  
+  .search-input {
+    border: none;
+    outline: none;
+    width: 100%;
+    font-size: 14px;
+  }
+  
+  .modal-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .modal {
+    background-color: white;
+    padding: 20px;
+    border-radius: 8px;
+    width: 400px;
+    max-width: 90%;
+  }
+  
+  .modal-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    border-bottom: 1px solid #ddd;
+    margin-bottom: 20px;
+    font-size: 14px;
+  }
+  
+  .modal-header h2 {
+    margin: 0;
+  }
+  
+  .modal-content label {
+    display: block;
+    margin-bottom: 5px;
+    font-weight: bold;
+  }
+  
+  .modal-content input {
+    width: calc(100% - 22px); /* Adjust width to include padding and border */
+    padding: 10px;
+    margin-bottom: 10px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    font-size: 14px;
+    box-sizing: border-box; /* Include padding and border in the element's total width and height */
+  }
+  
+  .modal-footer {
+    display: flex;
+    justify-content: flex-end;
+    border-top: 1px solid #ddd;
+    margin-top: 20px;
+    padding-top: 10px;
+  }
+  
+  .save-button {
+    background-color: #4caf50;
+    color: white;
+    padding: 10px 20px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 14px;
+    margin-right: 10px;
+  }
+  
+  .save-button:hover {
+    background-color: #45a049;
+  }
+  
+  .cancel-button {
+    background-color: #f44336;
+    color: white;
+    padding: 10px 20px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 14px;
+  }
+  
+  .cancel-button:hover {
+    background-color: #d32f2f;
+  }
+
