+++ gym-management-systems/home-page/src/components/Sidebar.js
@@ -1,0 +1,29 @@
+import React from 'react';
+import SidebarData from './SidebarData'; // Update this to use default import;
+import { useNavigate } from 'react-router-dom';
+
+function Sidebar() {
+    const navigate = useNavigate();
+
+    return (
+        <div className="Sidebar">
+            <ul className="SidebarList">
+                {SidebarData.map((val, key) => {
+                    return (
+                        <li
+                            key={key}
+                            className="row"
+                            id={window.location.pathname === val.link ? "active" : ""}
+                            onClick={() => navigate(val.link)}
+                        >
+                            <div id="icon">{val.icon}</div>
+                            <div id="title">{val.title}</div>
+                        </li>
+                    );
+                })}
+            </ul>
+        </div>
+    );
+}
+
+export default Sidebar;
